+++ sdk/typescript/src/dwallet-mpc/dkg.ts
@@ -239,7 +239,7 @@ async function launchDKGFirstRound(c: Config): Promise<DKGFirstRoundOutputResult
 				mutable: true,
 			}),
 			tx.pure.id(networkDecryptionKeyID),
-			tx.pure.u8(0),
+			tx.pure.u32(0),
 			emptyIKACoin,
 			tx.gas,
 		],
